Keyboard input added

+++ HeaderApp/NewTypes.h
@@ -10,7 +10,7 @@ typedef void(*prtFuncNext)(void);
 typedef struct _state
 {
     uint8_t counter;
-    uint8_t elements;
+    int16_t elements;
     uint8_t state;
     uint16_t bat;
     uint8_t character;
